requirements.txt
streamlit
google-generativeai
pillow
numpy
scipy
plotly
